Fixes automatic rom patching for gui and cli (path problem)
The ips/ups patching only worked form cli when called from the same
folder as the rom. It is because the rom path was never taken into
account, so it was a relative path, i.e. your pwd had to be in the rom
folder for patching to work. Now it should always work, I hope.