Add JACK audio support
This allows use of the JACK audio backend. JACK support
will be built by cmake if found, or the --with-jack option
to the configure script.

JACK audio will set the sample rate for audio equal to
the jack server's sample rate. It also connects to the first
two physical output ports for playback.