4 Player Adapter + NstDatabase.xml
I overlooked the NstDatabase.xml when adding options for users to manually assign adapters. If (dbpresent) is true, it would select the most suited adapter, and would not allow users to manually assign the adapter. I moved some code around, and added a bit, so everything should function correctly now even if NstDatabase.xml is present in system folder or not.