FLTK: License FLTKUI codebase under the 3-Clause BSD License
Since all FLTKUI related code is either written from scratch for the
upcoming release or taken from "The Jolly Good Reference Frontend", I
have chosen to relicense to align with the JGRF sources. This
codebase as a whole is still released under the GPLv2+, but the FLTKUI
portions may be useful for other JG frontends that may pop up in the
future and may need to link against non-GPL sources.