FLTK: Default to "Timer" sync mode
The world has changed a lot in recent years. VSync is going to give a
superior result if you run your monitor at a multiple of 60Hz on
X11 or Windows, but macOS and Wayland pose problems even then. Timer
is going to handle Wayland better at any refresh rate, and X11 and
Windows better at refresh rates that are not a multiple of 60. The
out-of-the-box experience is what people judge the project on, and
with Wayland becoming the default for popular distros, this is just
the safer choice.