README updates
Although it is unfortunate, macOS support for the FLTK port is still
not up to par with Linux and BSD. Changes in recent versions of macOS
seem to completely break frame pacing in FLTK, and seemingly also Qt
after some basic testing. It could be that some workarounds are
necessary, since SDL has no problem here, but it could also be the
long term effects of Apple hating OpenGL and leaving it to die on the
vine. Until this can be sorted out, macOS can't really be considered
fully supported, even with all of the effort put into it recently.